// Frontend/src/app/model/movie.ts
export class Movie {

    movieId: number | undefined;
    movieName: string | undefined;
    theatreName: string | undefined;
    totalSeats: number | undefined;

    // constructor(movieId: number, movieName: string, theatreName: string, totalSeats: number) {
    //     this.movieId = movieId;
    //     this.movieName = movieName;
    //     this.theatreName = theatreName;
    //     this.totalSeats = totalSeats;
    // }

}